requirements.txt
py-clob-client>=0.17.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.6.0
httpx>=0.27.0
websockets>=12.0
aiofiles>=23.2.1
python-multipart>=0.0.9
eth-account>=0.10.0